backend/requirements.txt
# SignLink 后端依赖
# Python版本要求: 3.8+
# 更新时间: 2025年11月 - 修复依赖兼容性问题
# 主要更新:
# - 更新MediaPipe到0.10.21版本
# - 更新NumPy到1.26.4版本
# - 保持TensorFlow 2.17.1版本（与README一致）

# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 机器学习
tensorflow==2.17.1
scikit-learn==1.3.2

# 计算机视觉
opencv-python==4.8.1.78
mediapipe==0.10.21
Pillow==10.1.0

# 数据处理
numpy==1.26.4

# HTTP请求和文件处理
python-multipart==0.0.6
aiofiles==23.2.0

# 异步支持
asyncio-mqtt==0.16.1

# 日志和监控
loguru==0.7.2

# 开发工具（可选）
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.11.0
# isort==5.12.0
# flake8==6.1.0
